requirements.txt
aiogram==3.14.0
python-dotenv==1.0.1
openai==1.55.0
requests==2.32.3